+++ xknx/secure/__init__.py
@@ -1,9 +1,10 @@
 """Classes for handling KNX IP Secure."""
 from .keyring import Keyring, load_key_ring
-from .util import sha256_hash
+from .util import bytes_xor, sha256_hash
 
 __all__ = [
     "Keyring",
     "load_key_ring",
+    "bytes_xor",
     "sha256_hash",
 ]
